backend_proxy/requirements.txt
Flask
requests
python-dotenv
Flask-CORS
yfinance
pandas 